requirements.txt
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.10.1
